Synthwave Post Process Half-Working
Outline and Edge detection works flawlessly, only need to make the grid and it should be good to go. Also, need to decide whether or not the filter should be applied to the character as well or if it should keep its default look or have a different filter applied to it.