Fedora Ability Working
Reworked the Projectile Parent to handle collision and therefore bounces.

Reworked the Damageable BPI and Damage System so that the Sandbox Character can handle the damage logic for the destroyable actors it spawns, eliminating the need to duplicate the Damage handling logic.

Need to change the Fedora's HP to the correct value once decided.