Wrath Passive Technically Working
The only thing missing is to reset it back to 0 on respawns and clean its code. Also found out that the Flame Prison is supposed to deal damage on spawn, which it does not do, so need to come back to it, it wasn't cleaned anyway I think. Synthwave gun anims should now be synced, projectile are now attributed their damage on spawn, so all Character changed and I couldn't be bothered to change Miku so I guess she does 0 damage now, to be fair she did 0.1 as far as I recall so that shouldn't bother her that much.