Metal Working
Everything should work all according to the GDD, the only thing that could be nit-picked is that the Wrath bar doesn't reset on death. Code need to be cleaned (A bit in Metal Character and the whole Flame Prison), will do tomorrow morning. Also, Miss Katty shouldn't be able to jump on her projectiles anymore.