Consecration Working
Spear now returns to the hand after its delay.

However, code is messy and I'm not entirely satisfied of how the Ability is coded, but I'll get to that when I clean it up.

Changing the way the health bar is handled make it so that health bar above the characters' head now longer updates correctly, I mean, it does, but it was originally built from values ranging from 0 to 1 and the new system uses the actual health, so will need to rework it urgently tomorrow morning.