Metal VFX Upgrade Half-Working
Improved the Flame Prison's VFX, but it's not perfect yet, distance field would be nice to have some of the big flames on the ground on the sphere as well to do a better transition. Speaking of those big flames, they should be moved out of the post process into a decal because they currently apply to the players and it looks quite bad. Although, even if it was pretty bad, it did give the illusion of flickering light onto the player, as if it was coming from the fire, could try to replicate that intentionally and better in the post process. Also, smoke and sparks where used in the Fire Slash I imported, it's a cool ideas so maybe I'll try to incorporate them in the VFX for the Flame Prison.