Metal Post Process Half-Fix
Now only need to fix the mirroring problem along the Z Axis on cross-sections. Also, by fixing it, it made the animation slightly worse, because now the flames on the walls no longer go up, the go outward, and it looked better for flames to go up, so might need to try to go back to that once the post process no longer has visual issues.