Point System Half-Working
Technically working but far from perfect, the same could be said of the overtime, it works but might be unprecise since it doesn't stop the timer. Might not be the cleanest code (not commented) and finally, the way it is handled is kind of counter intuitive, it adds points to the opposite team since its their Bass Area that is being captured, need to find the best spot in the code to make the swap.