Speed Pillar Working
Acceleration and Max Walk Speed were handled in the PreCMC Movement Function, hence why it didn't work, it now does but as for all movement related effect, the chance of it being correctly replicated are slim, will need to test it out. Pillar can now be rotated, allowing for placement on wall and ceilings.