Level Design Materials Working
Walls are now automatically brighter, deactivated the initialization of level blocks in construction script to be able to assign them barricade overlays. To assign a barricade overlay, click on the blueprint in the map, in its component select the static mesh and then assign the overlay there in its details.