Divine Wall VFX Half-Working
Needs to modified for the deploy animation. Also began to fix some of the side effects of switching to the Fever Bar system, notably the fact that failure is now equal to -100 instead of 0 and that the combo can now go really quickly to values like 100, which was near impossible in the old system and therefore causes unbalance.