Orchestral Ultimate Behaviour Basis
Gives shield on take off along with a golden wave VFX, wings VFX implemented, can shoot while flying to charge the ultimate, plunge when clicking on rhythm on the next beat after being in the air for at least or else auto-plunge after 3s. Basic collision system on impact but will need to go back to it after making the passive, it also doesn't have an impact VFX yet. Should no longer be able to be stuck in the levitation state.