Team Spawn Working
Team Spawns now have 4 different spawn points and cycles through them. Had to branch before respawn so that it only executes on server, otherwise it spawned them twice and messed up the value of the spawn index. It is kind of a lazy fix but seems to be functional.

Also added print strings time a pawn possession occurs with the hope that it helps find out what causes the spectator bug, those prints are green when a possession should succeed and red when it fails.