Miss Katty Fix Working
Now gives ball on spawn and destroy all balls on death. Implementation is a bit wonky and still not cleaned up but it seems to work (see Begin Play in ElectroSwingCharacter Blueprint, all is handled there or near it).

Discovered that the balls could deal damage to corpse, causing a bunch of bugs, need to fix ideally.