Miss Katty Wireless Swing Half-Working
The new Wireless is half-working, it successfully tries to reach the target location before coming back to the player, however, the condition to validate the target is based on bounces for now, which causes infinite attempts when the target is in the air (when aiming at a player, a pillar or simply too far) will be changed once the visual feedback for the target is added, detection should ideally be based on the same collision as picking up the ball.

Fixed the bug where Katty corpses could be hit, the feet collision wasn't ignoring the Team A & B channel.