Wireless Swing Rework Working
Wireless Swing now has a target widget to indicate where it aims, the icon stays as long as balls move towards it. It is only displayed locally. When a new target is created the old one is destroyed.

Fixed minor bugs, the passive now does damage that takes into account the fever bar, the ball icons are now fully centered on the ball no matter the resolution, the previous method didn't work properly, using local size it now works perfectly.

Code is still getting messier and messier, didn't take the time to clean it while I made/modified it.