Grappling Chain Rework Working
Behaves like Silksong's Clawline, pull towards the end point no matter if it hits or not. Now has a range that can be changed and still has a slight velocity boost at the end for a smoother transition after having finished grappling. Potential idea now that I think about it : could instead be a velocity boost in the opposite direction, like how Silksong knocks you back after hitting an enemy with the Clawline.